Refactor components and update navigation logic
Reformatted `MudTextField` and `MudButton` properties in `EditFacultyMemberComponent.razor` for better readability. Set default value for `Id` parameter to an empty string. Changed `OnInitialized` to `OnInitializedAsync` and updated null check for `Id` to use `string.IsNullOrEmpty(Id)`. Removed `HandleSaveAsync` method.

In `FaultyMembersComponent.razor`, updated `NavigateToUpsertForm` to accept an object context instead of a `Guid` id, and now checks if the context is of type `FacultyMemberDto` to navigate using the `Id` property. Updated `OnClick` event handler for `MudButton` to pass the entire context instead of just the `Id`.