Add AcademicId property to FacultyMember entity
This update introduces the `AcademicId` property to the `FacultyMember` entity and related classes, enabling the management of academic identifiers. The UI component `EditFacultyMemberComponent.razor` has been updated to include a new text field for `AcademicId`. Additionally, the `FacultyMemberDto` and `FacultyMemberService` classes have been modified to support this new property. A migration has also been created to add the `AcademicId` column to the `FacultyMembers` table in the database.